refactor(toast): sfc to tsx
